Extract out linked offer logic
