+++ crates/chia-sdk-driver/src/clear_signing.rs
@@ -3,6 +3,7 @@ mod delegated_spend;
 mod facts;
 mod inner_spend;
 mod issuance;
+mod linked_offer;
 mod memos;
 mod parsed_spend;
 mod requested_payments;
@@ -17,6 +18,7 @@ pub use delegated_spend::*;
 pub use facts::*;
 pub use inner_spend::*;
 pub use issuance::*;
+pub use linked_offer::*;
 pub use memos::*;
 pub use parsed_spend::*;
 pub use requested_payments::*;
